+++ ts/src/core/ixBuilders/OnboardTraderDelegated/index.ts
@@ -1,0 +1,11 @@
+export {
+  getOnboardTraderDelegatedCodec,
+  getOnboardTraderDelegatedDecoder,
+  getOnboardTraderDelegatedEncoder,
+} from "./codec";
+export { buildOnboardTraderDelegatedIx } from "./ix";
+export {
+  type OnboardTraderDelegatedAccounts,
+  type OnboardTraderDelegatedIx,
+  type OnboardTraderDelegatedParams,
+} from "./types";